test(schema-form): split nested-composition suites by behavior and make case counts exact
