+++ packages/canard/schema-form/src/core/__tests__/BranchStrategy.composition.setValue.test.ts
@@ -2,78 +2,25 @@ import { describe, expect, it } from 'vitest';
 
 import { delay } from '@winglet/common-utils';
 
-import { SetValueOption, nodeFromJsonSchema } from '@/schema-form/core';
-import type { JsonSchema } from '@/schema-form/types';
+import { nodeFromJsonSchema } from '@/schema-form/core';
 
 import type { ObjectNode } from '../nodes/ObjectNode';
+import {
+  type CompositionCase,
+  createSchema,
+  expressDefault,
+  standardDefault,
+} from './BranchStrategy.composition.fixtures';
 
-type CompositionScope = 'oneOf' | 'anyOf';
-type CompositionCase = [outer: CompositionScope, inner: CompositionScope];
-
-const compositionCases: CompositionCase[] = [
+describe.each([
   ['oneOf', 'oneOf'],
   ['oneOf', 'anyOf'],
   ['anyOf', 'oneOf'],
   ['anyOf', 'anyOf'],
-];
-
-const createSchema = (
-  outerScope: CompositionScope,
-  innerScope: CompositionScope,
-): JsonSchema => ({
-  type: 'object',
-  properties: {
-    enabled: { type: 'boolean', default: true },
-  },
-  [outerScope]: [
-    {
-      '&if': './enabled === true',
-      properties: {
-        config: {
-          type: 'object',
-          properties: {
-            mode: {
-              type: 'string',
-              enum: ['standard', 'express'],
-              default: 'standard',
-            },
-          },
-          [innerScope]: [
-            {
-              '&if': "./mode === 'standard'",
-              properties: {
-                cost: { type: 'number', default: 5.99 },
-                days: { type: 'number', default: 7 },
-              },
-            },
-            {
-              '&if': "./mode === 'express'",
-              properties: {
-                expressCost: { type: 'number', default: 15.99 },
-                hours: { type: 'number', default: 24 },
-              },
-            },
-          ],
-        },
-      },
-    },
-  ],
-});
-
-const standardDefault = {
-  enabled: true,
-  config: { mode: 'standard', cost: 5.99, days: 7 },
-};
-
-const expressDefault = {
-  enabled: true,
-  config: { mode: 'express', expressCost: 15.99, hours: 24 },
-};
-
-describe('BranchStrategy nested composition - external value injection', () => {
-  it.each(compositionCases)(
-    '%s → %s preserves explicit values injected before initial settlement',
-    async (outerScope, innerScope) => {
+] as const satisfies readonly CompositionCase[])(
+  'BranchStrategy nested composition %s → %s - setValue injection',
+  (outerScope, innerScope) => {
+    it('preserves explicit values injected before initial settlement', async () => {
       const node = nodeFromJsonSchema({
         jsonSchema: createSchema(outerScope, innerScope),
         onChange: () => {},
@@ -89,12 +36,9 @@ describe('BranchStrategy nested composition - external value injection', () => {
         enabled: true,
         config: { mode: 'standard', cost: 12.5, days: 3 },
       });
-    },
-  );
+    });
 
-  it.each(compositionCases)(
-    '%s → %s fills nested defaults for a partial overwrite',
-    async (outerScope, innerScope) => {
+    it('fills nested defaults for a partial overwrite', async () => {
       const node = nodeFromJsonSchema({
         jsonSchema: createSchema(outerScope, innerScope),
         onChange: () => {},
@@ -105,12 +49,9 @@ describe('BranchStrategy nested composition - external value injection', () => {
       await delay();
 
       expect(node.value).toEqual(standardDefault);
-    },
-  );
+    });
 
-  it.each(compositionCases)(
-    '%s → %s applies only the latest synchronous pre-settlement injection',
-    async (outerScope, innerScope) => {
+    it('applies only the latest synchronous pre-settlement injection', async () => {
       const node = nodeFromJsonSchema({
         jsonSchema: createSchema(outerScope, innerScope),
         onChange: () => {},
@@ -130,12 +71,9 @@ describe('BranchStrategy nested composition - external value injection', () => {
         enabled: true,
         config: { mode: 'express', expressCost: 25, hours: 8 },
       });
-    },
-  );
+    });
 
-  it.each(compositionCases)(
-    '%s → %s restores the active object when an overwrite omits it',
-    async (outerScope, innerScope) => {
+    it('restores the active object when an overwrite omits it', async () => {
       const node = nodeFromJsonSchema({
         jsonSchema: createSchema(outerScope, innerScope),
         onChange: () => {},
@@ -146,12 +84,9 @@ describe('BranchStrategy nested composition - external value injection', () => {
       await delay();
 
       expect(node.value).toEqual(standardDefault);
-    },
-  );
+    });
 
-  it.each(compositionCases)(
-    '%s → %s fills defaults for a partial non-default inner branch',
-    async (outerScope, innerScope) => {
+    it('fills defaults for a partial non-default inner branch', async () => {
       const node = nodeFromJsonSchema({
         jsonSchema: createSchema(outerScope, innerScope),
         onChange: () => {},
@@ -162,12 +97,9 @@ describe('BranchStrategy nested composition - external value injection', () => {
       await delay();
 
       expect(node.value).toEqual(expressDefault);
-    },
-  );
+    });
 
-  it.each(compositionCases)(
-    '%s → %s restores defaults after external deactivation and reactivation',
-    async (outerScope, innerScope) => {
+    it('restores defaults after external deactivation and reactivation', async () => {
       const node = nodeFromJsonSchema({
         jsonSchema: createSchema(outerScope, innerScope),
         onChange: () => {},
@@ -182,51 +114,9 @@ describe('BranchStrategy nested composition - external value injection', () => {
       await delay();
 
       expect(node.value).toEqual(standardDefault);
-    },
-  );
+    });
 
-  it.each(compositionCases)(
-    '%s → %s preserves an explicit constructor defaultValue',
-    async (outerScope, innerScope) => {
-      const node = nodeFromJsonSchema({
-        jsonSchema: createSchema(outerScope, innerScope),
-        defaultValue: {
-          enabled: true,
-          config: { mode: 'standard', cost: 8.5, days: 2 },
-        },
-        onChange: () => {},
-      }) as ObjectNode;
-
-      await delay();
-
-      expect(node.value).toEqual({
-        enabled: true,
-        config: { mode: 'standard', cost: 8.5, days: 2 },
-      });
-    },
-  );
-
-  it.each(compositionCases)(
-    '%s → %s fills nested defaults for a partial constructor defaultValue',
-    async (outerScope, innerScope) => {
-      const node = nodeFromJsonSchema({
-        jsonSchema: createSchema(outerScope, innerScope),
-        defaultValue: {
-          enabled: true,
-          config: { mode: 'express' },
-        },
-        onChange: () => {},
-      }) as ObjectNode;
-
-      await delay();
-
-      expect(node.value).toEqual(expressDefault);
-    },
-  );
-
-  it.each(compositionCases)(
-    '%s → %s filters fields from an inactive inner branch',
-    async (outerScope, innerScope) => {
+    it('filters fields from an inactive inner branch', async () => {
       const node = nodeFromJsonSchema({
         jsonSchema: createSchema(outerScope, innerScope),
         onChange: () => {},
@@ -249,53 +139,6 @@ describe('BranchStrategy nested composition - external value injection', () => {
         enabled: true,
         config: { mode: 'express', expressCost: 20, hours: 12 },
       });
-    },
-  );
-
-  it.each(compositionCases)(
-    '%s → %s preserves root fields during a nested Merge',
-    async (outerScope, innerScope) => {
-      const node = nodeFromJsonSchema({
-        jsonSchema: createSchema(outerScope, innerScope),
-        onChange: () => {},
-      }) as ObjectNode;
-
-      await delay();
-      node.setValue(
-        { config: { mode: 'standard', cost: 10, days: 1 } },
-        SetValueOption.Merge,
-      );
-      await delay();
-
-      expect(node.value).toEqual({
-        enabled: true,
-        config: { mode: 'standard', cost: 10, days: 1 },
-      });
-    },
-  );
-
-  it.each(compositionCases)(
-    '%s → %s preserves nested values during a selector-only Merge',
-    async (outerScope, innerScope) => {
-      const node = nodeFromJsonSchema({
-        jsonSchema: createSchema(outerScope, innerScope),
-        onChange: () => {},
-      }) as ObjectNode;
-
-      await delay();
-      node.setValue({
-        enabled: true,
-        config: { mode: 'standard', cost: 10, days: 1 },
-      });
-      await delay();
-
-      node.setValue({ enabled: true }, SetValueOption.Merge);
-      await delay();
-
-      expect(node.value).toEqual({
-        enabled: true,
-        config: { mode: 'standard', cost: 10, days: 1 },
-      });
-    },
-  );
-});
+    });
+  },
+);
